integers should be longs in some situations

+++ ConnectingApps.Refit.OpenAI/Files/Response/Datum.cs
@@ -10,8 +10,8 @@ namespace ConnectingApps.Refit.OpenAI.Files.Response
         public string Id { get; set; } = null!;
         public string Purpose { get; set; } = null!;
         public string Filename { get; set; } = null!;
-        public int Bytes { get; set; }
-        public int CreatedAt { get; set; }
+        public long Bytes { get; set; }
+        public long CreatedAt { get; set; }
         public string Status { get; set; } = null!;
         public object StatusDetails { get; set; } = null!;
     }
