+++ ConnectingApps.Refit.OpenAI/Files/Response/FilePostResponse.cs
@@ -6,8 +6,8 @@
         public string Id { get; set; } = null!;
         public string Purpose { get; set; } = null!;
         public string Filename { get; set; } = null!;
-        public int Bytes { get; set; }
-        public int CreatedAt { get; set; }
+        public long Bytes { get; set; }
+        public long CreatedAt { get; set; }
         public string Status { get; set; } = null!;
         public string StatusDetails { get; set; } = null!;
     }
